pyproject.toml
[project]
name = "causeway"
dynamic = ["version"]
description = "Rule enforcement and learning for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "sqlite-vec>=0.1.7a2",  # 0.1.6 has broken aarch64 Linux wheel
    "numpy",
    "mcp",
    "pydantic-ai",
    "sentence-transformers",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "openai",
    "rich",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
causeway = "causeway.cli:main"
causeway-mcp = "causeway.mcp:run"
causeway-check = "causeway.hooks.check_rules:main"
causeway-learn = "causeway.learning_agent:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["causeway"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require API keys)",
    "slow: marks tests as slow (deselect with '-m not slow')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
